% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-data.R
\name{get_pop_df}
\alias{get_pop_df}
\title{Get a Sheffield population data frame}
\usage{
get_pop_df(df_name)
}
\arguments{
\item{df_name}{Data frame name (character string; e.g. "df_pop_la_age").}
}
\value{
Data frame.
}
\description{
\code{get_pop_df} returns the Sheffield population data frame specified in
its argument.
}
\details{
Call \code{list_all_pop_dfs} for a list of valid arguments. Call
\code{load_all_pop_dfs} to load all Sheffield population data frames and
browse for the data frame of interest.
}
\examples{
df_age <- get_pop_df("df_pop_la_age")

}
