% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-data.R
\name{list_all_pop_dfs}
\alias{list_all_pop_dfs}
\title{Get a list of the Sheffield population data frames available}
\usage{
list_all_pop_dfs()
}
\value{
Vector of characters.
}
\description{
\code{list_all_pop_dfs} lists the Sheffield population data frames to
predefined variable name.
}
\details{
Call \code{load_all_pop_dfs} to load all Sheffield population data frames.
}
\examples{
print(list_all_pop_dfs())

}
