% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lkup_ward_lac}
\alias{lkup_ward_lac}
\title{Lookup for Wards and LACs}
\format{
A data frame with four variables: \code{Ward Code}, \code{ca_name},
\code{ca_number} and \code{asc_name}.
}
\usage{
lkup_ward_lac
}
\description{
Four Sheffield Wards in each LAC (same boundaries as ASC Localities)
}
\keyword{datasets}
