% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-data.R
\name{load_all_pop_dfs}
\alias{load_all_pop_dfs}
\title{Load all of the Sheffield population data frames}
\usage{
load_all_pop_dfs()
}
\value{
none
}
\description{
\code{load_all_pop_dfs} loads all the Sheffield population data frames
to predefined variable names.
}
\details{
Call \code{list_all_pop_dfs} for a list of the variable names that will be
created as Sheffield population data frames.
}
\examples{
load_all_pop_dfs()

}
