% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-data.R
\name{load_pop_df}
\alias{load_pop_df}
\title{Load a Sheffield population data frame}
\usage{
load_pop_df(df_name)
}
\arguments{
\item{df_name}{Data frame name (character string; e.g. "df_pop_la_age").}
}
\value{
none
}
\description{
\code{load_pop_df} loads the Sheffield population data frame specified in
its argument to a variable with the same name.
}
\details{
Call \code{list_all_pop_dfs} for a list of valid arguments. Call
\code{load_all_pop_dfs} to load all Sheffield population data frames and
browse for the data frame of interest.
}
\examples{
load_pop_df("df_pop_la_age")

}
