% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_la_age}
\alias{pop_la_age}
\title{Sheffield population by year of age (2020).}
\format{
A data frame with two variables: \code{Age} and \code{Persons}.
}
\usage{
pop_la_age
}
\description{
Sheffield population estimates provided by the ONS.
}
\keyword{datasets}
