% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_la_age_female}
\alias{pop_la_age_female}
\title{City female population by age (2020).}
\format{
A data frame with two variables: \code{Age} and \code{Females}.
}
\usage{
pop_la_age_female
}
\description{
Sheffield population estimates provided by the ONS.
}
\keyword{datasets}
