% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_la_age_male}
\alias{pop_la_age_male}
\title{Sheffield male population by year of age (2020).}
\format{
A data frame with two variables: \code{Age} and \code{Males}.
}
\usage{
pop_la_age_male
}
\description{
Sheffield population estimates provided by the ONS.
}
\keyword{datasets}
