% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_la_median_age}
\alias{pop_la_median_age}
\title{Median age of Sheffield population from 2001 to 2020.}
\format{
A data frame with two variables: \code{Year} and \code{Median age}.
}
\usage{
pop_la_median_age
}
\description{
Sheffield population estimates provided by the ONS.
}
\keyword{datasets}
