% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_la_yrs}
\alias{pop_la_yrs}
\title{City population from 2001 to 2020.}
\format{
A data frame with two variables: \code{Year} and \code{Population}.
}
\usage{
pop_la_yrs
}
\description{
Sheffield population estimates provided by the ONS.
}
\keyword{datasets}
