% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_la_yrs_age_gender}
\alias{pop_la_yrs_age_gender}
\title{Sheffield population by age and gender from 2000 to 2020.}
\format{
A data frame with four variables: \code{Gender}, \code{Age},
\code{Year} and \code{Persons}.
}
\usage{
pop_la_yrs_age_gender
}
\description{
Sheffield population estimates provided by the ONS.
}
\keyword{datasets}
