% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_lac_age_female}
\alias{pop_lac_age_female}
\title{LAC female population by age (2020).}
\format{
A data frame with five variables: \code{ca_name},
\code{ca_number}, \code{asc_name}, \code{Age} and \code{Females}.
}
\usage{
pop_lac_age_female
}
\description{
Local Area Committee female population aggregated from ONS Ward data. LACs
have the same boundaries as ASC (Adult Social Care) localities.
}
\keyword{datasets}
