% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_lac_age_male}
\alias{pop_lac_age_male}
\title{LAC male population by age (2020).}
\format{
A data frame with five variables: \code{ca_name},
\code{ca_number}, \code{asc_name}, \code{Age} and \code{Males}.
}
\usage{
pop_lac_age_male
}
\description{
Local Area Committee male population aggregated from ONS Ward data. LACs have
the same boundaries as ASC (Adult Social Care) localities.
}
\keyword{datasets}
