% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_lac_yrs}
\alias{pop_lac_yrs}
\title{LAC population between 2001 and 2020.}
\format{
A data frame with five variables: \code{ca_name}, \code{ca_number},
\code{asc_name}, \code{Year} and \code{Persons}.
}
\usage{
pop_lac_yrs
}
\description{
Sheffield population estimates provided by the ONS.
}
\keyword{datasets}
