% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_lsoa_age_band}
\alias{pop_lsoa_age_band}
\title{LSOA population by fifteen year age bands.}
\format{
A data frame with four variables: \code{LSOA Code}, \code{LSOA Name},
\code{Age band} and \code{Population}.
}
\usage{
pop_lsoa_age_band
}
\description{
Sheffield population provided by the ONS. 2020 mid-year estimates.
}
\keyword{datasets}
