% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_nhood_age_female}
\alias{pop_nhood_age_female}
\title{Neighbourhood female population by age (2020).}
\format{
A data frame with four variables: \code{nhood_code},
\code{nhood_name}, \code{Age} and \code{Females}.
}
\usage{
pop_nhood_age_female
}
\description{
Sheffield neighbourhood population estimates provided by Dale.
}
\keyword{datasets}
