% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_nhood_yrs_age}
\alias{pop_nhood_yrs_age}
\title{Neighbourhood population by age between 2017 and 2020.}
\format{
A data frame with four variables: \code{nhood_code},
\code{nhood_name}, \code{Year}, \code{Age} and \code{Persons}.

A data frame with four variables: \code{nhood_code},
\code{nhood_name}, \code{Year}, \code{Age}, \code{Gender} and \code{Persons}.
}
\usage{
pop_nhood_yrs_age

pop_nhood_yrs_age
}
\description{
Sheffield neighbourhood population estimates provided by Dale.

Sheffield neighbourhood population estimates provided by Dale.
}
\keyword{datasets}
