% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_ward_age}
\alias{pop_ward_age}
\title{Ward population by year of age (2020).}
\format{
A data frame with four variables: \code{Ward Code},
\code{Ward Name}, \code{Age} and \code{Persons}.
}
\usage{
pop_ward_age
}
\description{
Sheffield population estimates provided by the ONS.
}
\keyword{datasets}
