% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_ward_age_female}
\alias{pop_ward_age_female}
\title{Ward female population by age (2020).}
\format{
A data frame with four variables: \code{Ward Code},
\code{Ward Name}, \code{Age} and \code{Females}.
}
\usage{
pop_ward_age_female
}
\description{
Sheffield population estimates provided by the ONS.
}
\keyword{datasets}
