% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_ward_age_male}
\alias{pop_ward_age_male}
\title{Ward male population by age (2020).}
\format{
A data frame with four variables: \code{Ward Code},
\code{Ward Name}, \code{Age} and \code{Males}.
}
\usage{
pop_ward_age_male
}
\description{
Sheffield population estimates provided by the ONS.
}
\keyword{datasets}
