% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_ward_yrs}
\alias{pop_ward_yrs}
\title{Ward population between 2001 and 2020.}
\format{
A data frame with four variables: \code{Ward Code},
\code{Ward Name}, \code{Year} and \code{Persons}.
}
\usage{
pop_ward_yrs
}
\description{
Sheffield population estimates provided by the ONS.
}
\keyword{datasets}
