% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_ward_yrs_age_gender}
\alias{pop_ward_yrs_age_gender}
\title{Ward population by age and gender between 2001 and 2020.}
\format{
A data frame with six variables: \code{Ward Code}, \code{Ward Name},
\code{Year}, \code{Age}, \code{Gender} and \code{Persons}.
}
\usage{
pop_ward_yrs_age_gender
}
\description{
Sheffield population estimates provided by the ONS.
}
\keyword{datasets}
