% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsheff.R
\docType{package}
\name{popsheff}
\alias{popsheff}
\title{popsheff: A package of Sheffield population data frames}
\description{
The popsheff package provides functions to get or load individual data
frames. Plus functions to list or load all data frames.
}
\section{Author}{

Laurie Platt \email{laurie.platt@sheffield.gov.uk}
}

\section{See Also}{

\url{https://github.com/scc-pi/popsheff}
}

